Sales contact create  problem solved 02/08/2022

+++ database/migrations/2022_07_25_094737_create_sales_contract_items_table.php
@@ -15,7 +15,7 @@ class CreateSalesContractItemsTable extends Migration
     {
         Schema::create('sales_contract_items', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('sales_contract_id')->nullable();
+            $table->unsignedInteger('sales_contract_id')->nullable();
             $table->string('consignment_type')->nullable();
             $table->string('hs_code')->nullable();
             $table->string('processing_type')->nullable();
